Создала еще один модуль data и подключила его в main
